can choose where state is saved; move CSS to separate file
I've moved the CSS for styling writemaths areas to a separate file.

In order to allow several writemaths areas on one page, I've added a third argument to the WriteMaths constructor for defining the path an area's state should be saved to in localStorage. If not given, it won't be saved.