preview of maths above caret
for even more instant feedback, when you're editing a bit of maths a little box pops up above the input area showing what it'll look like.
It should make a good guess at where unclosed maths environments end -- it stops when it gets to a word of two letters or more.